Refactor and outsource model specific methods in a new class "InferenceEngine".
